requirements.txt
# Django==3.1.2
# django-crispy-forms==1.9.2
# loguru==0.5.3
# six==1.15.0
# django-csp==3.7
# django-feature-policy==3.5.0
# django-axes==5.8.0
# python-dotenv==0.15.0

Django==4.0.3
django-crispy-forms==1.14.0
loguru==0.6.0
six==1.16.0
django-csp==3.7
django-permissions-policy==4.9.0
django-axes==5.31.0
python-dotenv==0.20.0